DXE-6532 Move datasources
