+++ pkg/providers/reportinggroups/provider.go
@@ -56,5 +56,7 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewCPCodeDataSource,
+		NewReportingGroupDataSource,
+		NewReportingGroupsDataSource,
 	}
 }
