+++ pkg/providers/reportinggroups/testdata/TestDataReportingGroupsGroup/reporting_group.tf
@@ -2,7 +2,7 @@ provider "akamai" {
   edgerc = "../../common/testutils/edgerc"
 }
 
-data "akamai_property_reportinggroups_group" "test" {
+data "akamai_reportinggroups_group" "test" {
   reporting_group_id = 12345
 }
