+++ pkg/providers/reportinggroups/testdata/TestDataReportingGroupsGroup/reporting_group_id_missing.tf
@@ -2,6 +2,6 @@ provider "akamai" {
   edgerc = "../../common/testutils/edgerc"
 }
 
-data "akamai_property_reportinggroups_groups" "test" {
+data "akamai_reportinggroups_group" "test" {
 }
