+++ pkg/providers/reportinggroups/testdata/TestDataReportingGroupsGroups/reporting_groups_with_contract_id.tf
@@ -2,7 +2,7 @@ provider "akamai" {
   edgerc = "../../common/testutils/edgerc"
 }
 
-data "akamai_property_reportinggroups_groups" "test" {
+data "akamai_reportinggroups_groups" "test" {
   contract_id = "ctr_456"
 }
